LocalServer/requirements.txt
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
pydantic>=1.10.0,<2.0.0
numpy>=1.24.0
joblib>=1.3.0
pytest>=7.4.0
httpx>=0.25.0
google-generativeai>=0.3.0

